adição do html ao php

+++ delete.php
@@ -1,19 +1,19 @@
-<body>
-    
-    
-    <?php
-require 'db.php';
+<?php
+require 'php/db.php';
 
-$id = $_GET['id'];
 
-$sql = "DELETE FROM livros WHERE id = ?";
-$stmt = $pdo->prepare($sql);
-$stmt->execute([$id]);
+if (isset($_GET['id'])) {
+    $id = $_GET['id'];
 
-header('Location: index.php');
+   
+    $sql = 'DELETE FROM livros WHERE id = ?';
+    $stmt = $pdo->prepare($sql);
+    $stmt->execute([$id]);
+
+   
+    header('Location: index.php');
+    exit();
+} else {
+    echo "ID inválido ou não fornecido!";
+}
 ?>
-
-
-<script src="js/scripts.js"></script>
-
-</body>
